add class Comment, changed other classes

+++ app/src/main/java/ru/netology/nework/domain/event/GetEventByIdUseCase.kt
@@ -2,7 +2,7 @@ package ru.netology.nework.domain.event
 
 class GetEventByIdUseCase(private val eventListRepository: EventListRepository) {
 
-    fun getEventById(id: Long): Event {
+    fun getEventById(id: Int): Event {
        return eventListRepository.getEventById(id)
     }
 }
