+++ app/src/main/java/ru/netology/nework/domain/job/JobListRepository.kt
@@ -4,6 +4,6 @@ interface JobListRepository {
     fun addJob(job: Job)
     fun deleteJob(job: Job)
     fun editJob(job: Job)
-    fun getJobById(id: Long): Job
+    fun getJobById(id: Int): Job
     fun getJobList() : List<Job>
 }
