+++ app/src/main/java/ru/netology/nework/domain/post/GetPostByIdUseCase.kt
@@ -2,7 +2,7 @@ package ru.netology.nework.domain.post
 
 class GetPostByIdUseCase(private val postListRepository: PostListRepository) {
 
-    fun getPostById(id: Long): Post {
+    fun getPostById(id: Int): Post {
         return postListRepository.getPostById(id)
     }
 }
