+++ app/src/main/java/ru/netology/nework/domain/post/PostListRepository.kt
@@ -5,7 +5,7 @@ interface PostListRepository {
     fun addPost(post: Post)
     fun deletePost(post: Post)
     fun editPost(post: Post)
-    fun getPostById(id: Long): Post
+    fun getPostById(id: Int): Post
     fun getPostList() : List<Post>
 
     //методы на лайки и поделиться + счетчики
